UF-XYZ - Domain refactoring
* Renames the SmsRequest model class to SendSmsRequest, for easier distinction in the messaging service
