+++ src/test/java/se/sundsvall/smssender/integration/linkmobility/LinkMobilityServiceTest.java
@@ -16,7 +16,7 @@ import org.springframework.http.HttpEntity;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.web.client.RestTemplate;
 
-import se.sundsvall.smssender.api.domain.SmsRequest;
+import se.sundsvall.smssender.api.model.SendSmsRequest;
 import se.sundsvall.smssender.integration.linkmobility.domain.LinkMobilityResponse;
 import se.sundsvall.smssender.integration.linkmobility.domain.ResponseStatus;
 
@@ -66,8 +66,8 @@ class LinkMobilityServiceTest {
         assertThat(sendSmsRequest.getUserData()).isEqualTo(request.getMessage());
     }
 
-    private SmsRequest validRequest() {
-        return SmsRequest.builder().withSender("sender")
+    private SendSmsRequest validRequest() {
+        return SendSmsRequest.builder().withSender("sender")
                 .withSender("sender")
                 .withMessage("message")
                 .withMobileNumber("+46701234567")
